Color_Detection/requirements.txt
opencv-python-headless
flask
gunicorn
pillow
numpy